net: phy: Re-apply 39194a40eef5a4771404ea04910b8216f8d2a065
Apparently, commit 39194a40eef5 ("Revert "net: phy: Correctly process
PHY_HALTED in phy_stop_machine()"") was reverted during the initial
synchronize codes for OxygenOS commit.

This was taken from the RenderZenith kernel. The only reason I am not
picking that commit is to make it clear that the original commit was
already in this tree. It's merely being re-applied.

https://github.com/EAS-Project/op3-oreo-kernel/commit/95691d6c78fce4065db07422d8cdca96b1334dc3

Fixes: 8cc3973a1385 ("Synchronize codes for OxygenOS 5.0.0")
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>